Move TemplateType from runtime to compiler, it is now an implementation detail and does not need to be exposed
