+++ jte/src/main/java/gg/jte/compiler/TemplateType.java
@@ -1,4 +1,4 @@
-package gg.jte.runtime;
+package gg.jte.compiler;
 
 public enum TemplateType {
     Template,
